database connection for 3 classes using postman get/push requests. update to map.tsx to improve performance

+++ backend/src/main/java/trinitytransit/backend/backend/repository/BusesRepository.java
@@ -1,7 +1,7 @@
 package trinitytransit.backend.backend.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
 import trinitytransit.backend.backend.entity.Buses;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BusesRepository extends JpaRepository<Buses, Integer> {
+public interface BusesRepository extends JpaRepository<Buses, Long> {
 }
